update tests to use helpers throughout
